refactor speed handling
Separate speed planning from gcode writing. For now, a fixed number of
layers is planned in advance. The goal is to slow down already before
the start of a very small layer. This will allow to smooth out speed
changes over multiple layers in a later commit.

Replace the list of points (of a planned path) with a list of
segments (new class GCodePathSegment). A segment stores not only the
XY position, but also the speed and extrusion width.

This will allow manipulations of the per-segment speed in a later
commit, to smooth out speed changes.

Rename forceMinimalLayerTime() --> enforceSpeedLimits(), which now
does all speed limits.