smooth speed changes
This fixes the messup often seen in the last layers of objects that
don't have a flat top. Slowdown for cooling now starts long before
reaching the layer which needs it.

While this doesn't fix the underlying problem (which is that the bot
actually doesn't follow flow changes immediately), it distributes the
error over a very large area, making it practically disappear.

It's working. But code needs cleanup & parameter & some debugging.
Integer speed resolution in mm/s is somewhat spoiling the fun.