increase resolution of speed
Convert integer speed to double in places where it matters.

The speed resolution as an integer in mm/s is too coarse. You can
actually hear the integer steps. And in the gcode we have mm/min.

The alternative would be to store the speed in microns/s and work in
fixed-point, but it would be messier in the code. Concerning
performance, this commit is actually a tiny (1%) improvement.